[API] Several enhancements to the python API.
Two parts to this change.

First, make the python bindings more OO, by encapsulating more
information in the objects. The objets now also expose method
so that information that is already kept in the object doesn't
need to be extracted to call the existing functions. The objects
also keep track of the resource that created them. So, instead
of calling:

  r = roles.create_role(resource, 'service', 'foo', 'bar',
      'host', 'cluster')

Now the call is much simpler:

  r = serviceObj.create_role('foo', 'bar', 'host')

Second, commands and configuration for services and roles are now
exposed. The new command object exposes only the OO interface.

The existing procedural interface is kept intact, so that existing
users don't break.