[API] Add python bindings for the CM calls.
. change the services / roles code to call into the '/cm' namespace
  when a cluster is not provided.
. add a new ClouderaManager object that encapsulates CM calls.

In the setup_cms() method, I chose to "manually" create the JSON request
instead of creating a proper python version of the ApiServiceSetupInfo
class. If/when we convert other calls (such as create service / role)
to take "setup info" objects, we may want to spend some time making
the code more generic.

Uploading a new license is not implemented yet.