[API] Add API resource object to python bindings.
This change wraps the generic Resource class in an API-specific
implementation that provides accessors to the API's top-level
methods (such as cluster and host methods).

This more closely mimics the structure of the Java resources,
and also avoids having to manually propagate the resource object
to different functions.

Also change systest/test_api.py to use the new OO style.