[API] Add endpoints to retrieve role logs.
Three new endpoints: the main log file (used "full" as the name to match the UI),
stdout and stderr. Available now are just the raw contents; no tailing, searching,
or anything fancy.

Added a new exception mapping to cover the case when the agent does not yet have
a heartbeat; return a SERVICE_UNAVAILABLE in that case (according to spec, that
means "try again later", which is the case here).

The logic to figure out log URLs was abstracted into new methods LogController.

Added some basic unit tests, but there's not much that can be unit tested without
an agent running.