[API] Be more lenient with host json in python bindings.
I've hit an issue a few times where the data returned from SCM
contains all the host's fields except for the ip address. This
would cause an error in the python bindings since it requires
all the r/w fields in the response.

This generally happens when you bring up SCM and it gets an
"optimized" heartbeat from the agent, and you query it before
SCM can fetch all the information about the agent. So it's
not something that would usually happen, but...