[API] Add EnableHA command for HDFS services.
Add the command implementation and the API version of the custom command
arguments structure for the enable HA command. The unit tests mostly cover
the data validation path, not the enable HA command itself.

Note that currently the command needs some baby sitting under the covers
to function correctly; that will eventually be fixed in the command itself,
at which point the API invocation should just work.