[API] Move command endpoints to top level.
This avoids replication of "/commands/{id}" and "/commands/{id}/abort"
under every resource that has commands associated with it. It also gets
rid of a lot of duplicated code, and also code we were using to filter
commands based on the context.

The "/commands" resource-specific endpoint to list active commands still
exists (at least in some resources, didn't check all), and will remain
that way to make it simpler to query resource-specific commands.