[API] Add "init stand-by NameNode" command.
This command is currently part of the "enable HA" workflow, but is
also exposed in the UI. It allows people to "manually" enable HA
by executing the steps of the workflow by themselves.

I also added some missing APIs to the python bindings, which I needed
to test this, and clarify the failover API, which made me waste a few
cycles to figure out what was wrong.