Add enter and exit maintenance mode for services, hosts and roles.
This change adds commands to enter and exit maintenance mode for services, roles
and hosts. I followed the pattern established by the code in clusters.py. At the
same time, I expose the concept of a synchronous command, which I also need in
the systest I wrote that actually uses this stuff.