[py] Add Hive replication arguments to python bindings.
This change not only adds those arguments, but also re-works
the replication API in the bindings a little bit so that it's
more easily extended, although it's probably not very easy to use.

I also renamed ApiPeer to ApiCmPeer, to match the CM name, and
made some small changes to the code that translates JSON data
so that it better copes with the data returned by the replication
URLs.