[py] Provide more descriptive attributes for API types.
This change introduces a new way to define attributes for
API types. With the extra metadata provided by the attributes,
a few features are possible:

. Easily nest types (with full serde support) without having
  to override common methods.

. Types now disallow setting unknown or read-only properties.

. Better sharing of code in the deserialization and construction
  paths.

A new method (_get_attributes) was added to allow for types that
reference themselves (such as ApiCommand).

One drawback of the change is that custom constructors must have
only keyword arguments (aside from resource_root). This greatly
simplified the implementation of BaseApiObject.from_json_dic().

As part of the change, I also changed the following:

. replication types now properly handle nested types
. 'displayName' property was missing from ApiService
. some time-related properties are now properly parsed
  into datetime objects. Note that this change may
  break existing scripts that expect time fields to be
  strings.