[services] Fix typo in the arg name of enable_jt_ha api
This change fixes the parameter name from "new_jt_host" to "new_jt_host_id"
in the enable_jt_ha api.

Jayita reported this when testing. As she is the only one
that has systest for this, I am going to create tarball with
this change and let her confirm the fix.