OPSAPS-16642. Expose API to fetch replication history.
Need this API to fix OPSAPS-16642, for which the replication systest
requires this API to fetch the "full" view of replication result.

Note that we only expose this API rather than additionally also adding a
"view" parameter to the get_replication_schedule(s) API. This is consistent
with the snapshot policy API and also prevents at least python API users
from fetching replication schedules with "full" view and potentially
overloading the server.