[py] Add update_user API call.
To support updating the state of a user in a generic way, add an update_user
API call that takes an ApiUser object and sends it back to the server to
update the referenced user.

Also clarified the behaviour of grant/revoke_admin_role in the presence
of multiple possible user roles.