Bump current API version in api_client.py to 10 for CM 5.4
This is an old commit that I forgot to push. It initially bumped
the API version to 10 in both setup.py and api_client.py, but
since I wrote it, the setup.py code was updated by another commit.
So this commit just updates API_CURRENT_VERSION in setup.py.