[api] OPSAPS-23455 add option to associated case, upload yarn application support bundle to Cloudera
Added a new function create_yarn_application_diagnostics_bundle to yarn service.
The new function adds the support for new arguments ticketNumber and comments.

The signature of the function create_yarn_application_diagnostics_bundle is not
compatible with the existing function collect_yarn_application_diagnostics and
that is why, I added a new function and marked the older one as deprecated.