OPSAPS-27837 Fix python API wrappers that don't mark list return values as lists
In get_yarn_application_attributes and get_impala_query_attributes, we expect
the return to be a list but did not set so. This cause json failures when they
are invoked.

Fix:
Set ret_is_list to True.

Testing Done:
Updated test_yarn and test_impala which was using wrong response json.