[api] OPSAPS-31699 Add SSL context argument to ApiResource
Starting in Python 2.7, SSL certificate validation is turned on by
default (yay!), but means that if CM is using an SSL certificate not
signed by a public CA, Python API calls will fail. To allow specifying
a local CA certificate, we need to allow API users to pass in a custom
SSL context. This change adds the SSL context argument to ApiResource.
Because HTTPSHandler only supports the context argument in Python 2.7.9
and above, this feature is only available in those versions.

Testing:
- Added an example script which connects to CM using a custom CA cert
- Tested against my cluster which has HTTPS turned on