[API] OPSAPS-48697 Don't update latest API version for 5.16.0
Latest CM API version is v30 that was released with CM 6.0.

We didn't update CM API version in 5.15 and 5.16 and it remains
at v19 for those releases.

Hence latest API version must not be updated on releasing
5.16.0.

Tests:
- Tested 5.16.0 API docs are listed on releases page
and latest CM API version links are not updated
using local jekyll server
http://bankim-desktop.pa.cloudera.com:4000/cm_api/docs/releases/